Extract MCP output observability helpers
Move bounded output-token and cache-status metric helpers out of mcp-server so #687 Wave 5 can shrink the server facade without changing exported helper contracts.

Constraint: Existing tests import these helpers from src/mcp-server, so mcp-server must re-export the moved functions.

Rejected: moving journal summaries in the same slice | journal behavior has separate tests and should remain a distinct extraction.

Confidence: high

Scope-risk: narrow

Directive: Keep future MCP Wave 5 slices behavior-preserving and re-export old helper paths until downstream imports migrate.

Tested: npm test -- tests/metrics/tool-output-observability.test.ts tests/tools/journal.test.ts --runInBand

Tested: npm run build

Tested: npm run lint -- --quiet

Tested: npm run lint:tier

Tested: git diff --check

Not-tested: full repository test suite
